trying to fix the Quad_List.
Why was quad_list not converted to a X_D3DPT_TRIANGLELIST
X_D3DPT_QUADSTRIP is converted to X_D3DPT_TRIANGLESTRIP.. and both are Xbox only !!!

Now only missing the 2e part of the quad display.
a quad are 2 polygons connected to each other to each other.

This still not complete fix the quad_list. but it will not crash.