XTL_EmuIDirect3DDevice8_BlockUntilVerticalBlank need to wait for virticalBlank.. this is not a segagt thing.. it on every game.. so don't use the vsync check there..
function TIDirectSound.GetTime
is crashing.. not correct way to get time.. 
now its disabled.

function IsEmuHandle(hFile: {xboxkrnl.}HANDLE): Boolean; inline;
the >= 0 was before <> -1.. is also not correct.... changed the check to <> -2
Rayman results in -2 for not emufiles.. 


Result.. 
bladeII gives again the intro flash. Movie still not start.
something with buffer size.

Rayman is working 
and Ghost Masters starts with intro movie.

This is the first time all 3 are running.. still not perfect, but before it was that 1 was running and other 3 where broken.